Query

+++ app/Controllers/Http/AdminsController.ts
@@ -1,8 +1,10 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import Admin from 'App/Models/Admin'
+import MailDelivery from 'App/Services/Kafka/MailDelivery'
+import StoreValidator from 'App/Validators/Admins/CRUD/StoreValidator'
 
 export default class AdminsController {
-  public async index({ response }: HttpContextContract) {
+  public async index({ response, request }: HttpContextContract) {
     try {
       const admins = await Admin.query().select('id', 'full_name', 'email')
       return admins
@@ -14,9 +16,11 @@ export default class AdminsController {
   }
 
   public async store({ request, response }: HttpContextContract) {
+    await request.validate(StoreValidator)
     try {
       const data = request.all()
       const admin = await Admin.create(data)
+      await new MailDelivery().send(admin, { admin }, 'newAdmin', 'Boas vindas')
       return admin
     } catch (err) {
       response
@@ -25,22 +29,39 @@ export default class AdminsController {
     }
   }
 
-  public async show({ params, response }: HttpContextContract) {
-    const admin = await Admin.find(params.id)
-    if (!admin) {
-      return response.notFound({
-        error: { message: 'Oops, esse usuário não existe ou foi apagado' },
-      })
-    }
-    return {
-      id: admin.id,
-      name: admin.fullName,
-      email: admin.email,
-      createdAt: admin.createdAt,
+  public async show({ params, auth, response }: HttpContextContract) {
+    try {
+      if (Number(auth.user!.id) !== Number(params.id)) {
+        return response.forbidden({
+          error: { message: 'Oops, você não tem permissão para realizar essa operacação.' },
+        })
+      }
+
+      const admin = await Admin.find(params.id)
+      if (!admin) {
+        return response.notFound({
+          error: { message: 'Oops, esse usuário não existe ou foi apagado' },
+        })
+      }
+      return {
+        id: admin.id,
+        name: admin.fullName,
+        email: admin.email,
+        createdAt: admin.createdAt,
+      }
+    } catch (err) {
+      response
+        .status(err.status)
+        .send({ error: { message: 'Oops, algo deu ao buscar os seus dados' } })
     }
   }
 
-  public async update({ params, request, response }: HttpContextContract) {
+  public async update({ params, auth, request, response }: HttpContextContract) {
+    if (Number(auth.user!.id) !== Number(params.id)) {
+      return response.forbidden({
+        error: { message: 'Oops, você não tem permissão para realizar essa operacação.' },
+      })
+    }
     const admin = await Admin.find(params.id)
     if (!admin) {
       return response.notFound({
@@ -53,13 +74,24 @@ export default class AdminsController {
     return admin
   }
 
-  public async destroy({ params, response }: HttpContextContract) {
-    const admin = await Admin.find(params.id)
-    if (!admin) {
-      return response.notFound({
-        error: { message: 'Oops, esse usuario não existe ou já foi apagado' },
-      })
+  public async destroy({ params, auth, response }: HttpContextContract) {
+    try {
+      if (Number(auth.user!.id) !== Number(params.id)) {
+        return response.forbidden({
+          error: { message: 'Oops, você não tem permissão para realizar essa operacação.' },
+        })
+      }
+      const admin = await Admin.find(params.id)
+      if (!admin) {
+        return response.notFound({
+          error: { message: 'Oops, esse usuario não existe ou já foi apagado' },
+        })
+      }
+      await admin.delete()
+    } catch (err) {
+      response
+        .status(err.status)
+        .send({ error: { message: 'Oops, algo deu ao apagar os seus dados' } })
     }
-    await admin.delete()
   }
 }
